rename package in omdb-client api module
