+++ movie-service/src/main/java/com/garret/movies/service/config/mappings/impl/OmdbMovieToMovieDtoConfigurer.java
@@ -1,6 +1,6 @@
 package com.garret.movies.service.config.mappings.impl;
 
-import com.garret.movies.omdb.entity.OmdbMovie;
+import com.garret.movies.omdb.dto.OmdbMovie;
 import com.garret.movies.service.config.mappings.ModelMapperConfigurer;
 import com.garret.movies.service.dto.ActorDto;
 import com.garret.movies.service.dto.CountryDto;
