+++ web/src/main/java/com/garret/movies/batch/ShortMovieReader.java
@@ -1,7 +1,7 @@
 package com.garret.movies.batch;
 
 import com.garret.movies.omdb.api.OmdbClient;
-import com.garret.movies.omdb.entity.ShortMovie;
+import com.garret.movies.omdb.dto.ShortMovie;
 import org.springframework.batch.core.annotation.BeforeStep;
 import org.springframework.batch.core.configuration.annotation.StepScope;
 import org.springframework.batch.item.ItemReader;
