+++ web/src/test/java/com/garret/movies/batch/MovieDbWriterTest.java
@@ -1,6 +1,6 @@
 package com.garret.movies.batch;
 
-import com.garret.movies.omdb.entity.OmdbMovie;
+import com.garret.movies.omdb.dto.OmdbMovie;
 import com.garret.movies.service.api.MovieService;
 import com.garret.movies.service.dto.response.MovieDto;
 import org.junit.Before;
